Add embedded jobs spreadsheet and link to this spreadsheet
This commit adds an embed of my spreadsheet of job opportunities and
application status from Google Sheets. It also adds a link to this
spreadsheet. In addition, the spreadsheet and link only show up if I am
signed in.

Successfully adds conditional to show or hide embedded spreadsheet depending on whether I am signed in

Add link to spreadsheet

finished formatting link